refactor: move unstable crypto files
