+++ crypto/unstable_keystack.ts
@@ -8,8 +8,8 @@
  * @module
  */
 
-import { timingSafeEqual } from "../timing_safe_equal.ts";
-import * as base64url from "../../encoding/base64url.ts";
+import { timingSafeEqual } from "./timing_safe_equal.ts";
+import * as base64url from "../encoding/base64url.ts";
 
 /** Types of data that can be signed cryptographically. */
 export type Data = string | number[] | ArrayBuffer | Uint8Array;
@@ -74,7 +74,7 @@ async function compare(a: Data, b: Data): Promise<boolean> {
  *
  * @example
  * ```ts
- * import { KeyStack } from "https://deno.land/std@$STD_VERSION/crypto/keystack.ts";
+ * import { KeyStack } from "https://deno.land/std@$STD_VERSION/crypto/unstable_keystack.ts";
  *
  * const keyStack = new KeyStack(["hello", "world"]);
  * const digest = await keyStack.sign("some data");
